+++ src/VirtoCommerce.Platform.Core/Modularity/ContentFileTypes.cs
@@ -1,0 +1,13 @@
+namespace VirtoCommerce.Platform.Core.Modularity;
+
+/// <summary>
+/// File-asset kinds the modularity framework recognises. The web layer's
+/// <c>ContentFile.Type</c> property carries these values verbatim — they are
+/// kept as plain strings so the JSON contract stays stable across future
+/// C# renames.
+/// </summary>
+public static class ContentFileTypes
+{
+    public const string Script = "script";
+    public const string Style = "style";
+}
